+++ odoo/local-src/specific_product/models/__init__.py
@@ -7,3 +7,6 @@ from . import network
 from . import network_api
 from . import res_company
 from . import stock
+from . import product_uom_categ
+from . import product_template
+from . import product_uom