+++ odoo/local-src/specific_product/models/product_uom.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+# Copyright 2017 Camptocamp SA
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from odoo import fields, models
+
+
+class ProductUom(models.Model):
+    _inherit = 'product.uom'
+
+    recurring = fields.Boolean(
+        related='category_id.recurring',
+    )