+++ odoo/local-src/specific_sale/models/__init__.py
@@ -1,2 +1,3 @@
 from . import company
 from . import sale
+from . import sale_order_line