+++ odoo/local-src/specific_sale/wizard/__init__.py
@@ -1,1 +1,2 @@
 from . import sale_refusal
+from . import mrp_invoicing